auth-service/backend/requirements.txt
django==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
requests==2.22.0
gunicorn==20.0.4
psycopg2==2.8.6
djangorestframework-simplejwt==4.6.0
pyjwt==2.0.0